fix: (read msg) plugin build fails when installed from remote with Lazy
The plugin first needs to find project's
'base-folder', so that it can then locate the
folder where the 'lib*parser.so' should
be compiled.

To find this it uses Lua's 'debug.getinfo()',
which returns two kinds of paths,
"source" and "short_src". I shuold have used
the former, instead I used the ladder.

That's because when getting the path,
'debug.getinfo()' will truncate and prepend with '...'
the returned path if too long.

When the plugin gets installed from github
through lazy, it gets by default installed
in '.local/share/nvim/lazy/...':

now the full path is too long and it's unusable,
(has '...' prepended + is incomplete) which causes
a linking error since it cannot find the path.